Microservices in Java - Spring Data JPA + H2 Database - Refactor

+++ 0x04/JPA+H2_database/src/main/java/com/example/jpa_h2_demo/CustomerController.java
@@ -8,59 +8,78 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 import org.springframework.http.HttpStatus;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.hateoas.EntityModel;
+import org.springframework.hateoas.CollectionModel;
 
-import java.util.List;
+import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
+import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
+
+import org.springframework.data.domain.Pageable;
 
 @RestController
+@RequestMapping(path = "/customers")
 public class CustomerController {
-    @Autowired
-    private CustomerRepository customerRepository;
+    private final CustomerRepository customerRepository;
+    private final CustomerModelAssembler customerModelAssembler;
 
-    @PostMapping("/customers")
-    public Customer addCustomer(@RequestBody Customer customer) {
-        if (customer.getPhones() != null) {
-            for (Phone phone : customer.getPhones()) {
-                phone.setCustomer(customer);
-            }
-        }
-
-        if (customer.getAddresses() != null) {
-            for (Address address : customer.getAddresses()) {
-                address.setCustomer(customer);
-            }
-        }
-
-        return this.customerRepository.save(customer);
+    public CustomerController(
+        CustomerRepository customerRepository,
+        CustomerModelAssembler customerModelAssembler
+    ) {
+        this.customerRepository = customerRepository;
+        this.customerModelAssembler = customerModelAssembler;
     }
 
-    @GetMapping("/customers")
-    public List<Customer> findAllCustomers() {
-        return this.customerRepository.findAll();
+    @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
+    public EntityModel<Customer> addCustomer(@RequestBody Customer customer) {
+        return this.customerModelAssembler.toModel(this.customerRepository.save(customer));
     }
 
-    @GetMapping("/customers/{id}")
-    public Customer findCustomerById(@PathVariable("id") Long id) {
-        return this.customerRepository.findById(id).orElseThrow(
-            () -> new RuntimeException(
-                String.format("Connot find Customer: no customer with id=%d", id)
+    @GetMapping
+    public CollectionModel<EntityModel<Customer>> findAllCustomers(Pageable pageable) {
+        return CollectionModel.of(
+            this.customerModelAssembler.toCollectionModel(this.customerRepository.findAll(pageable)),
+            linkTo(methodOn(CustomerController.class).findAllCustomers(pageable)).withSelfRel()
+        );
+    }
+
+    @GetMapping(path = "/{id}")
+    public EntityModel<Customer> findCustomerById(@PathVariable("id") Long id) {
+        return this.customerModelAssembler.toModel(
+            this.customerRepository.findById(id).orElseThrow(
+                () -> new CustomerNotFoundException(id)
             )
         );
     }
 
-    @DeleteMapping("/customers/{id}")
+    @DeleteMapping(path = "/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void removeCustomerById(@PathVariable("id") Long id) {
+        if (!this.customerRepository.existsById(id)) {
+            throw new CustomerNotFoundException(id);
+        }
+
         this.customerRepository.deleteById(id);
     }
 
-    @PutMapping("/customers/{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void updateCustomerById(@PathVariable("id") Long id, @RequestBody Customer customer) {
-        customer.setId(id);
-        this.customerRepository.save(customer);
+    @PutMapping(path = "/{id}")
+    public EntityModel<Customer> updateCustomerById(
+        @PathVariable("id") Long id,
+        @RequestBody Customer customer
+    ) {
+        Customer customerToUpdate = this.customerRepository.findById(id).orElseThrow(
+            () -> new CustomerNotFoundException(id)
+        );
+
+        customerToUpdate.setPhones(customer.getPhones());
+        customerToUpdate.setAddresses(customer.getAddresses());
+        customerToUpdate.setData(customer.getData());
+
+        return this.customerModelAssembler.toModel(this.customerRepository.save(customerToUpdate));
     }
 }
